+++ app/src/layers/components/CustomLayer.jsx
@@ -4,6 +4,7 @@ import classnames from 'classnames';
 import CustomLayerStyles from 'styles/components/map/custom-layer.scss';
 import MapFormStyles from 'styles/components/map/form.scss';
 import ButtonStyles from 'styles/components/button.scss';
+import { CUSTOM_LAYERS_SUBTYPES } from 'constants';
 
 class CustomLayer extends Component {
 
@@ -12,7 +13,8 @@ class CustomLayer extends Component {
 
     this.state = {
       name: 'Layer name',
-      description: ''
+      description: '',
+      subtype: 'geojson'
     };
   }
 
@@ -44,13 +46,68 @@ class CustomLayer extends Component {
     }
 
     return (
-      <div className={CustomLayerStyles.customLayer} >
+      <div className={CustomLayerStyles.customLayer}>
         <form
           className={classnames(MapFormStyles.form, CustomLayerStyles.uploadForm)}
           onSubmit={e => this.onSubmit(e)}
         >
-          <div className={CustomLayerStyles.column} >
-            <div className={CustomLayerStyles.row} >
+          <div className={CustomLayerStyles.column}>
+            <div className={CustomLayerStyles.row}>
+              <label className={MapFormStyles.fieldName} htmlFor="name">Type</label>
+              <div className={MapFormStyles.radioGroup}>
+                <input
+                  type="radio"
+                  name="subtype"
+                  id={CUSTOM_LAYERS_SUBTYPES.geojson}
+                  value={CUSTOM_LAYERS_SUBTYPES.geojson}
+                  checked={this.state.subtype === CUSTOM_LAYERS_SUBTYPES.geojson}
+                  onChange={e => this.onChange(e.currentTarget)}
+                />
+                <label htmlFor={CUSTOM_LAYERS_SUBTYPES.geojson}>
+                  GeoJSON
+                </label>
+              </div>
+              <div className={MapFormStyles.help}>
+                A simple vector format that can be produced by various GIS such as ArcGIS or QGIS.
+                You first need to upload the file, using a service such as Dropbox or Github.
+              </div>
+              <div className={MapFormStyles.radioGroup}>
+                <input
+                  type="radio"
+                  name="subtype"
+                  id={CUSTOM_LAYERS_SUBTYPES.raster}
+                  value={CUSTOM_LAYERS_SUBTYPES.raster}
+                  checked={this.state.subtype === CUSTOM_LAYERS_SUBTYPES.raster}
+                  onChange={e => this.onChange(e.currentTarget)}
+                />
+                <label htmlFor={CUSTOM_LAYERS_SUBTYPES.raster}>
+                  Raster
+                </label>
+              </div>
+              <div className={MapFormStyles.help}>
+                Use this option for data that can be consumed as tiles (images), such as
+                satellite imagery.
+              </div>
+              <div className={MapFormStyles.radioGroup}>
+                <input
+                  type="radio"
+                  name="subtype"
+                  id="wms"
+                  value="wms"
+                  checked={this.state.subtype === 'wms'}
+                  onChange={e => this.onChange(e.currentTarget)}
+                />
+                <label htmlFor="wms">
+                  WMS
+                </label>
+              </div>
+              <div className={MapFormStyles.help}>
+                Web Map Service (WMS) following the OGC standard and capable of serving raster tiles from a GIS database.
+              </div>
+            </div>
+          </div>
+          <div className={CustomLayerStyles.column}>
+            <div className={CustomLayerStyles.row}>
               <label className={MapFormStyles.fieldName} htmlFor="name" >name</label>
               <input
                 className={MapFormStyles.textInput}
@@ -61,22 +118,24 @@ class CustomLayer extends Component {
                 required
               />
             </div>
-
-            <div className={CustomLayerStyles.row} >
+            <div className={CustomLayerStyles.row}>
               <label className={MapFormStyles.fieldName} htmlFor="url" >url</label>
               <input
                 className={MapFormStyles.textInput}
                 name="url"
-                placeholder="Insert a link to a .geojson layer file"
                 type="text"
                 onChange={e => this.onChange(e.currentTarget)}
                 required
+                placeholder={{
+                  geojson: 'Link to a GeoJSON file',
+                  raster: 'Raster tiles URL template with XYZ parameters',
+                  wms: 'WMS server root'
+                }[this.state.subtype]
+                }
               />
             </div>
-          </div>
-          <div className={CustomLayerStyles.column} >
-            <div className={CustomLayerStyles.row} >
-              <label className={MapFormStyles.fieldName} htmlFor="description" >description</label>
+            <div className={CustomLayerStyles.row}>
+              <label className={MapFormStyles.fieldName} htmlFor="description" >description (optional)</label>
               <textarea
                 className={MapFormStyles.textarea}
                 name="description"
@@ -87,11 +146,14 @@ class CustomLayer extends Component {
             </div>
           </div>
 
-          <div className={CustomLayerStyles.row} >
-            {this.props.error &&
-            <span className={CustomLayerStyles.submitError} > Whoops! Something went wrong. </span>
+          <div className={CustomLayerStyles.row}>
+            {this.props.error !== null &&
+            <span className={CustomLayerStyles.submitError}>
+              Whoops! Something went wrong.<br />
+              Please check if the custom layer type is properly selected or if the data you provided is correct.
+            </span>
             }
-            <div className={CustomLayerStyles.submitContainer} >
+            <div className={CustomLayerStyles.submitContainer}>
               <input
                 className={classnames(ButtonStyles.button, ButtonStyles._filled,
                   ButtonStyles._big, CustomLayerStyles.submitButton)}